requirements.txt
fastapi[all]>=0.111.0
cryptography>=38.0.4
uvicorn>=0.30.0
pydantic>=2.8.2
pydantic-settings>=2.3.4
SQLAlchemy>=2.0.31
SQLAlchemy-Utils>=0.41.2
python-dotenv>=1.0.1
shortuuid>=1.0.11
psycopg>=3.1.20


